week_05/Assignment_14/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langgraph
langchain-tavily
faiss-cpu
python-dotenv